# API/openapi/API_v2.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Tranga v2",
    "version": "2.0"
  },
  "paths": {
...
    "/v2/Search/{MangaConnectorName}/{Query}": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Initiate a search for a API.Schema.MangaContext.Manga on API.Controllers.DTOs.MangaConnector with searchTerm",
        "parameters": [
          {
            "name": "MangaConnectorName",
            "in": "path",
            "description": "API.Controllers.DTOs.MangaConnector.Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Query",
            "in": "path",
            "description": "searchTerm",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API.Controllers.DTOs.MinimalManga exert of API.Schema.MangaContext.Manga",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MinimalManga"
                  }
                }
              }
            }
          },
          "404": {
            "description": "API.Controllers.DTOs.MangaConnector with Name not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "API.Controllers.DTOs.MangaConnector with Name is disabled"
          }
        }
      }
    },
    "/v2/Search/{MangaConnectorName}/Manga": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Returns full API.Schema.MangaContext.Manga detail from a API.Controllers.DTOs.MangaConnector by its site ID, without saving to the database",
        "parameters": [
          {
            "name": "MangaConnectorName",
            "in": "path",
            "description": "API.Controllers.DTOs.MangaConnector.Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ConnectorMangaId",
            "in": "query",
            "description": "The manga's ID on the connector site",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full API.Controllers.DTOs.Manga detail",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/Manga"
                }
              }
            }
          },
          "404": {
            "description": "Manga not found on connector",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AltTitle": {
        "required": [
          "language",
          "title"
        ],
        "type": "object",
        "properties": {
          "language": {
            "minLength": 1,
            "type": "string",
            "description": "Language of the Title"
          },
          "title": {
            "minLength": 1,
            "type": "string",
            "description": "Title"
          }
        },
        "additionalProperties": false,
        "description": "API.Schema.MangaContext.AltTitle DTO"
      },
      "Author": {
        "required": [
          "key",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Author."
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "The API.Schema.MangaContext.Author DTO"
      },
...
      "Link": {
        "required": [
          "key",
          "provider",
          "url"
        ],
        "type": "object",
        "properties": {
          "provider": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Provider"
          },
          "url": {
            "minLength": 1,
            "type": "string",
            "description": "Url"
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "API.Schema.MangaContext.Link DTO"
      },
      "Manga": {
        "required": [
          "altTitles",
          "authors",
          "description",
          "ignoreChaptersBefore",
          "key",
          "links",
          "mangaConnectorIds",
          "name",
          "releaseStatus",
          "tags"
        ],
        "type": "object",
        "properties": {
          "ignoreChaptersBefore": {
            "type": "number",
            "description": "Chapter cutoff for Downloads (Chapters before this will not be downloaded)",
            "format": "float"
          },
          "year": {
            "type": "integer",
            "description": "Release Year",
            "format": "int32",
            "nullable": true
          },
          "originalLanguage": {
            "type": "string",
            "description": "Release Language",
            "nullable": true
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Author"
            },
            "description": "Author-names"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Manga Tags"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "description": "Links for more Metadata"
          },
          "altTitles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AltTitle"
            },
            "description": "Alt Titles of Manga"
          },
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Manga"
          },
          "description": {
            "minLength": 1,
            "type": "string",
            "description": "Description of the Manga"
          },
          "releaseStatus": {
            "$ref": "#/components/schemas/MangaReleaseStatus"
          },
          "mangaConnectorIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MangaMangaConnectorId"
            },
            "description": "Ids of the Manga on MangaConnectors"
          },
          "coverUrl": {
            "type": "string",
            "description": "External cover image URL from the connector",
            "nullable": true
          },
          "fileLibraryId": {
            "type": "string",
            "description": "Id of the Library the Manga gets downloaded to",
            "nullable": true
          },
          "language": {
            "type": "string",
            "description": "Content Language (e.g. translation language)",
            "nullable": true
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "API.Schema.MangaContext.Manga DTO"
      },
...
      "MangaMangaConnectorId": {
        "required": [
          "foreignKey",
          "key",
          "mangaConnectorName",
          "objId",
          "useForDownload"
        ],
        "type": "object",
        "properties": {
          "mangaConnectorName": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Connector"
          },
          "foreignKey": {
            "minLength": 1,
            "type": "string",
            "description": "Key of the referenced DTO"
          },
          "objId": {
            "minLength": 1,
            "type": "string",
            "description": "ID of the Object on the connector site"
          },
          "websiteUrl": {
            "type": "string",
            "description": "Website Link for reference, if any",
            "nullable": true
          },
          "useForDownload": {
            "type": "boolean",
            "description": "Whether this Link is used for downloads"
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "API.Controllers.DTOs.MangaConnectorId`1 DTO"
      },
      "MangaReleaseStatus": {
        "enum": [
          "Continuing",
          "Completed",
          "OnHiatus",
          "Cancelled",
          "Unreleased"
        ],
        "type": "string"
      },
...
      "MinimalManga": {
        "required": [
          "description",
          "key",
          "mangaConnectorIds",
          "name",
          "releaseStatus"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Manga"
          },
          "description": {
            "minLength": 1,
            "type": "string",
            "description": "Description of the Manga"
          },
          "releaseStatus": {
            "$ref": "#/components/schemas/MangaReleaseStatus"
          },
          "mangaConnectorIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MangaMangaConnectorId"
            },
            "description": "Ids of the Manga on MangaConnectors"
          },
          "coverUrl": {
            "type": "string",
            "description": "External cover image URL from the connector",
            "nullable": true
          },
          "fileLibraryId": {
            "type": "string",
            "description": "Id of the Library the Manga gets downloaded to",
            "nullable": true
          },
          "language": {
            "type": "string",
            "description": "Content Language (e.g. translation language)",
            "nullable": true
          },
          "key": {
            "maxLength": 64,
            "minLength": 16,
            "type": "string",
            "description": "Unique Identifier of the DTO"
          }
        },
        "additionalProperties": false,
        "description": "Shortened Version of API.Controllers.DTOs.Manga"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}